Move configs and builtins into their own directories
