+++ tpot/config/regressor_light.py
@@ -98,7 +98,7 @@ regressor_config_dict_light = {
     'sklearn.preprocessing.StandardScaler': {
     },
 
-    'tpot.built_in_operators.ZeroCount': {
+    'tpot.builtins.ZeroCount': {
     },
 
     # Selectors
